New Mission: Operation Rescue
This is my first completed mission. It's not perfect, and it could probably do with some polishing. However, I want to make this live for feedback purposes. Players are able to complete it as well as fail. I would like opinions on difficulty, AI setting and the general idea. Hopefully with this creation published, I can achieve that.