Redundancy to Aeon M2 if attackers killed quickly
I had a game where the map wouldn't expand after the attackers were killed.  Adding in logs to this revealed all the M3Attackers were returning true for :IsDead().
Adding in the code to force the next mission objective to run if all attackers are dead resolved the issue and meant the map expanded.